Major rewrite of OM, annotations
Split actual object management to ObjectRepository class, OM is purely a convenience layer now
Annotations never include files to parse; much safer, less memory intensive, makes it possible to prime cache with an already primed cache
Use TokenReflection to parse imports in AnnotationReader, for consistency

+++ OM.php
@@ -35,20 +35,6 @@
 
 namespace Fossil;
 
-use Fossil\Exceptions\NoSuchClassException;
-
-class InstanceWrapper {
-    private $reflClass;
-
-    public function __construct($fqcn) {
-        $this->reflClass = new \ReflectionClass($fqcn);
-    }
-
-    public function create() {
-        return $this->reflClass->newInstanceArgs(func_get_args());
-    }
-}
-
 /**
  * The object manager class
  * 
@@ -68,36 +54,13 @@ class InstanceWrapper {
  */
 class OM {
     /**
-     * Maps type to current instance
-     * 
-     * @var array
+     *
+     * @var ObjectRepository
      */
-    private static $instances = array();
-    private static $extensionClasses = array();
-    private static $classMap = array();
-    private static $scannedClasses = array();
-    private static $instanceWrappers = array();
+    private static $objectRepo;
 
     private static $startupTime = 0;
 
-    /**
-     * Maps type to potential provider name
-     * 
-     * Provider keys:
-     * ['fqcn']         => Fully Qualified Class Name
-     * ['default']      => Whether this provider is to be used by default
-     * ['takesContext'] => Whether this provider has a constructor which
-     *                     takes the previous provider instance for context
-     * 
-     * @var array
-     */
-    private static $singletonClasses = array(
-        'FS' => array('default' => array('fqcn' => '\\Fossil\\Filesystem', 'takesContext' => false)),
-        'Annotations' => array('default' => array('fqcn' => '\\Fossil\\Annotations\\AnnotationManager', 'takesContext' => false)),
-        'Error' => array('default' => array('fqcn' => '\\Fossil\\ErrorManager', 'takesContext' => false)),
-        'Settings' => array('default' => array('fqcn' => '\\Fossil\\Settings', 'takesContext' => false))
-    );
-    private static $instancedClasses = array();
     private static $dirty = false;
     
     protected static function makeDirty() {
@@ -121,90 +84,18 @@ class OM {
         file_put_contents(self::FS()->tempDir() . D_S . '.quickstart.yml', yaml_emit($quickstart));
     }
 
-    private static function scanForSingletonObjects() {
-        $allClasses = self::Annotations()->getClassesWithAnnotation("F:Object");
-        foreach($allClasses as $class) {
-            if(isset(self::$scannedClasses[$class]))
-                continue;
-            self::$scannedClasses[$class] = true;
-
-            $annotations = self::Annotations()->getClassAnnotations($class, "F:Object");
-            foreach($annotations as $objAnno) {
-                $type = $objAnno->value ?: $objAnno->type;
-                if(!isset(self::$singletonClasses[$type]))
-                    self::$singletonClasses[$type] = array();
-                self::$singletonClasses[$type][$objAnno->name] = array('fqcn' => '\\' . $class, 'takesContext' => $objAnno->takesContext);
-            }
-        }
-    }
-
-    private static function scanForInstancedObjects() {
-        $allClasses = self::Annotations()->getClassesWithAnnotation("F:Instanced");
-
-        foreach($allClasses as $class) {
-            if(isset(self::$scannedClasses[$class]))
-                continue;
-            self::$scannedClasses[$class] = true;
-
-            $reflClass = new \ReflectionClass($class);
-            // Skip the class if it's abstract
-            if($reflClass->isAbstract())
-                continue;
-            $annotations = self::Annotations()->getClassAnnotations($class, "F:Instanced");
-            foreach($annotations as $objAnno) {
-                if(!isset($objAnno->type)) {
-                    // Check the class's namespace
-                    $namespace = $reflClass->getNamespaceName();
-                    $type = substr($namespace, strrpos($namespace, '\\')+1);
-                } else {
-                    $type = $objAnno->type;
-                }
-
-                if(isset($objAnno->name)) {
-                    $name = $objAnno->name;
-                } elseif(isset($objAnno->value)) {
-                    $name = $objAnno->value;
-                } else {
-                    $name = substr($class, strrpos($class, "\\")+1);
-                }
-                // Normalize the name before storing it
-                $name = ucfirst(strtolower($name));
-
-                if(!isset(self::$instancedClasses[$type]))
-                    self::$instancedClasses[$type] = array();
-                self::$instancedClasses[$type][$name] = $class;
-            }
-        }
-    }
-
-    private static function scanForObjects($root) {
-        // Get all php files below this directory, excluding libs and static
-        $files = OM::FS()->sourceFiles($root);
-        
-        foreach($files as $file) {
-            try
-            {
-                include_once($file);
-            }
-            catch(\Exception $e)
-            {
-            }
-        }
-        self::Annotations()->updateAnnotations();
-        self::scanForSingletonObjects();
-        self::scanForInstancedObjects();
-        self::$extensionClasses = self::Annotations()->getClassesWithAnnotation("F:ExtensionClass");
-    }
-
-    public static function setup() {
+    public static function setup($considerMtimes = true) {
         self::$startupTime = microtime(true);
-        self::Error()->init(E_ALL | E_STRICT);
+        // Start with a basic object repo
+        self::$objectRepo = new ObjectRepository();
+        
+//        self::Error()->init(E_ALL | E_STRICT);
         
         // Load the basic settings from 'quickstart.yml'
         if(!file_exists(self::FS()->tempDir() . D_S . '.quickstart.yml'))
             return;
         // Ignore the quickstart if it's newer than settings.yml
-        if(file_exists(self::FS()->execDir() . D_S . 'settings.yml')) {
+        if($considerMtimes && file_exists(self::FS()->execDir() . D_S . 'settings.yml')) {
             if(filemtime(self::FS()->tempDir() . D_S . '.quickstart.yml') <
                filemtime(self::FS()->execDir() . D_S . 'settings.yml')) {
                 unlink(self::FS()->tempDir() . D_S . '.quickstart.yml');
@@ -218,13 +109,14 @@ class OM {
         // If we have settings, grab the cache
         if(isset($basics['cache'])) {
             // Bypass the setTypeInstance function so as not to dirty the context
-            self::$instances['Cache'] = new $basics['cache']['fqcn']($basics['cache']['config']);
+            self::$objectRepo->setSingleton("Cache", new $basics['cache']['fqcn']($basics['cache']['config']));
         }
     }
 
     public static function getRuntime() {
         return microtime(true) - self::$startupTime;
     }
+    
     /**
      * Initialize the object manager without cache
      * 
@@ -252,7 +144,7 @@ class OM {
      * 
      *  @return bool Whether cached state could be loaded
      */
-    public static function cachedInit() {
+    public static function cachedInit($considerMtimes = true) {
         if(!self::has("Cache"))
             return false;
         
@@ -261,22 +153,22 @@ class OM {
             return false;
         
         // Check the mtimes of basic classes
-        if(self::getFossilMtime() > $cachedData['mtime'])
-            return false;
+        if($considerMtimes)
+            if(self::getFossilMtime() > $cachedData['mtime'])
+                return false;
         
-        self::$singletonClasses = $cachedData['singleton'];
-        self::$instancedClasses = $cachedData['instanced'];
+        self::$objectRepo = $cachedData['objects'];
         
         // Get the compiler, to set up the namespace path
         self::Compiler()->registerAutoloadPath();
         
-        self::$classMap = $cachedData['classMap'];
-        self::$instances['Annotations'] = new Annotations\AnnotationManager($cachedData['annotations']);
+        self::Annotations()->loadFromCache($cachedData['annotations']);
         
         // Then after loading the plugin manager etc, check mtimes again, just in case
         self::Plugins()->loadEnabledPlugins();
-        if(self::getFossilMtime() > $cachedData['mtime'])
-            return false;
+        if($considerMtimes)
+            if(self::getFossilMtime() > $cachedData['mtime'])
+                return false;
         
         // Finally, register plugins with the ORM
         self::ORM()->registerPaths();
@@ -303,178 +195,53 @@ class OM {
     public static function primeCache() {
         // Regular functionality:
         // Scan local namespace for objects
-        self::scanForObjects(self::FS()->fossilRoot());
-        foreach(OM::FS()->roots(false) as $root)
-            self::scanForObjects($root);
+        self::$objectRepo->scanForObjects(false);
         // Load settings up, set up drivers
         self::Compiler()->registerAutoloadPath();
-        self::get('Cache');
+        self::$objectRepo->getSingleton('Cache');
         self::ORM()->ensureSchema(true);
         // Register plugins
         // TODO: Move to auto-plugin loader
         OM::Plugins()->loadEnabledPlugins();
         // Scan plugin namespaces for objects
-        foreach(self::FS()->pluginRoots() as $root)
-            self::scanForObjects($root);
+        self::$objectRepo->scanForObjects(true);
         // Do compilation
         $tempClassMap = OM::Compiler()->bootstrap();
         OM::Compiler()->setClassMap($tempClassMap);
-        self::$classMap = OM::Compiler()->compileAll();
+        self::$objectRepo->setClassMap(OM::Compiler()->compileAll());
     }
 
     public static function saveCache() {
         $cachedData = array();
-        $cachedData['singleton'] = self::$singletonClasses;
-        $cachedData['instanced'] = self::$instancedClasses;
-        $cachedData['classMap'] = self::$classMap;
-        $cachedData['annotations'] = self::Annotations()->gatherAnnotations();
+        $cachedData['objects'] = self::$objectRepo;
+        $cachedData['annotations'] = self::Annotations()->dumpForCache();
         $cachedData['mtime'] = self::getFossilMtime();
         
         OM::Cache()->set("fossil_state", $cachedData);
     }
-    
-    /**
-     * Select which provider to use for a type
-     * 
-     * Creates an instance of the new provider immediately,
-     * passing the old provider to the new one as required
-     * 
-     * Pre-condition: $name must be known as a $type
-     * 
-     * @param string $type The type which is being managed
-     * @param string $name The name of the provider to use
-     * @return void
-     */
-    public static function select($type, $name, $fqcn = null) {
-        if($fqcn) {
-            self::setTypeInstance($type, new $fqcn);
-            return;
-        }
-
-        // Pre-condition: $name must be known as a $type
-        // Post-condition: There will be an instance of $type stored in the OM
-        $newInstance = NULL;
-        $oldInstance = NULL;
-        // Remove the current instance before instantiating the new one
-        if(isset(self::$instances[$type])) {
-            $oldInstance = self::$instances[$type];
-            self::$instances[$type] = NULL;
-        }
-        // Then create the new instance, giving it context if it wants it
-        $typeInfo = self::$singletonClasses[$type][$name];
-        $class = $typeInfo['fqcn'];
-        if(isset(self::$classMap[$class]))
-            $class = self::$classMap[$class];
-        if($typeInfo['takesContext']) {
-            $newInstance = new $class($oldInstance);
-        } else {
-            $newInstance = new $class;
-        }
-        self::setTypeInstance($type, $newInstance);
-    }
-
-    private static function setTypeInstance($type, $instance) {
-        // We only care about cache when setting the dirty status
-        if($type == "Cache") {
-            self::makeDirty();
-        }
-
-        self::$instances[$type] = $instance;
-    }
-
-    private static function selectDefault($type) {
-        // First off, check that we know about this type
-        if(!isset(self::$singletonClasses[$type]) || !isset(self::$singletonClasses[$type]['default'])) {
-            $typeName = ObjectFactory::getObjectName($type);
-            if(!$typeName) {
-                throw new \Exception("No default object found for type $type");
-            }
-            self::select($type, $typeName);
-            return;
-        }
-        
-        // By default, just use the element named default
-        $typeInfo = self::$singletonClasses[$type]['default'];
-        $class = $typeInfo['fqcn'];
-        if(isset(self::$classMap[$class]))
-            $class = self::$classMap[$class];
-
-        // Instantiate it
-        try
-        {
-            $newInstance = new $class;
-        }
-        catch(\Fossil\Exceptions\SelectionChangedException $e) {
-            return;
-        }
-        // And store it
-        self::setTypeInstance($type, $newInstance);
-    }
-
-    private static function resolveInstanceClass($typeOrFqcn, $subtype = null) {
-        if($subtype) {
-            // Normalize the subtype before retrieval, as before storage
-            $subtype = ucfirst(strtolower($subtype));
-            if(!isset(self::$instancedClasses[$typeOrFqcn]) || !isset(self::$instancedClasses[$typeOrFqcn][$subtype]))
-                throw new NoSuchClassException($typeOrFqcn, $subtype);
-
-            $typeOrFqcn = self::$instancedClasses[$typeOrFqcn][$subtype];
-        }
-        if(isset(self::$classMap[$typeOrFqcn]))
-            return self::$classMap[$typeOrFqcn];
-        // Fallback, if none was found
-        return $typeOrFqcn;
-    }
 
     public static function obj($typeOrFqcn, $subtype = null) {
-        $actualClass = self::resolveInstanceClass($typeOrFqcn, $subtype);
-
-        if(!isset(self::$instanceWrappers[$actualClass]))
-            self::$instanceWrappers[$actualClass] = new InstanceWrapper($actualClass);
-
-        return self::$instanceWrappers[$actualClass];
+        return self::$objectRepo->getInstanceWrapper($typeOrFqcn, $subtype);
     }
     
     public static function getAllInstanced($type) {
-        if(isset(self::$instancedClasses[$type]))
-            return self::$instancedClasses[$type];
-        return array();
-    }
-
-    /**
-     * Get an object of the requested type
-     * 
-     * Instantiates the object if none yet exists.
-     * Throws an exception if the object manager doesn't
-     * know about the requested type
-     * 
-     * @param string $type The type to retrieve
-     * @return mixed An instance of the requested type
-     */
-    public static function get($type) {
-        if(!isset(self::$instances[$type])) {
-            self::selectDefault($type);
-        }
-        return self::$instances[$type];
+        return self::$objectRepo->getAllInstanceClasses($type);
     }
     
-    public static function getSpecific($type, $name) {
-        if(!isset(self::$singletonClasses[$type]))
-            throw new Exception("Unknown type: $type");
-        if(!isset(self::$singletonClasses[$type][$name]))
-            throw new Exception("Unknown class of type $type: $name");
-        return self::$singletonClasses[$type][$name];
+    public static function getSpecificSingleton($type, $name) {
+        return self::$objectRepo->getSpecificSingleton($type, $name);
     }
 
-    public static function getAll($type) {
-        return self::$singletonClasses[$type];
+    public static function getAllSingletons($type) {
+        return self::$objectRepo->getAllSingletonClasses($type);
     }
 
-    public static function getBaseObjects() {
+    public static function getObjectsToCompile() {
         // TODO: Determine whether to only compile in-use objects or not
         $classList = array();
 
-        foreach(self::$singletonClasses as $classArr) {
+        $singletonClasses = self::$objectRepo->getAllSingletonClasses();
+        foreach($singletonClasses as $classArr) {
             foreach($classArr as $classDat) {
                 // Don't return already compiled objects - that way, madness lies
                 if(strpos($classDat['fqcn'], '\\Fossil\\Compiled') !== 0) {
@@ -482,7 +249,8 @@ class OM {
                 }
             }
         }
-        foreach(self::$instancedClasses as $typeArr) {
+        $instancedClasses = self::$objectRepo->getAllInstanceClasses();
+        foreach($instancedClasses as $typeArr) {
             foreach($typeArr as $type)
                 $classList[] = $type;
         }
@@ -491,7 +259,7 @@ class OM {
     }
 
     public static function getExtensionClasses() {
-        return self::$extensionClasses;
+        return self::$objectRepo->getExtensionClasses();
     }
 
     /**
@@ -501,7 +269,7 @@ class OM {
      * @return string TRUE if the type has been instantiated
      */
     public static function has($type) {
-        return isset(self::$instances[$type]);
+        return self::$objectRepo->hasSingleton($type);
     }
 
     /**
@@ -513,14 +281,18 @@ class OM {
      * @param string $type The name of the type to check for
      * @param mixed $name The name of the provider to check for, or NULL
      */
-    public static function knows($type, $name=NULL) {
-        if(!isset(self::$singletonClasses[$type]))
-            return false;
-        if($name && !isset(self::$singletonClasses[$type][$name]))
-            return false;
-        return true;
+    public static function knows($type, $name = null) {
+        return self::$objectRepo->knowsSingleton($type, $name);
     }
 
+    public static function select($type, $name) {
+        self::$objectRepo->selectSingleton($type, $name);
+    }
+    
+    public static function setSingleton($type, $instance) {
+        self::$objectRepo->setSingleton($type, $instance);
+    }
+    
     /**
      * Static magic method calls
      * 
@@ -532,14 +304,17 @@ class OM {
      * @return mixed A singleton instance of the requested type
      */
     public static function __callStatic($type, $args) {
-        if(isset(self::$instances[$type]) || isset(self::$singletonClasses[$type])) {
-            $obj = self::get($type);
+        try
+        {
+            $obj = self::$objectRepo->getSingleton($type);
             if(count($args) > 0)
                 return call_user_func_array(array($obj, 'get'), $args);
             else
-                return self::get($type);
+                return $obj;
         }
-        throw new \BadMethodCallException("Method '$type' does not exist");
+        catch(Exception $e) {
+            throw new \BadMethodCallException("Method '$type' does not exist");
+        }
     }
     
     protected static $appNS;
@@ -575,5 +350,15 @@ class OM {
             $name .= "_" . basename(self::$overlayNS);
         }
         return $name;
+    }
+    
+    public static function getFossilName() {
+        $name = "Fossil";
+        if(isset(self::$overlayNS))
+            return basename(self::$overlayNS);
+        else if(isset(self::$appNS))
+            return basename(self::$appNS);
+        else
+            return $name;
     }
 }
